Move account scoped lambda service methods to separate file
