+++ plugin-hrm-form/src/services/fetchProtectedApi.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2021-2023 Technology Matters
+ * Copyright (C) 2021-2026 Technology Matters
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
  * by the Free Software Foundation, either version 3 of the License, or
@@ -34,9 +34,10 @@ export class ProtectedApiError extends ApiError {
 /**
  * Factored out function that handles a protected api call hosted in serverless toolkit.
  * Will throw Error if server responses with and http error code.
-
+ * TODO: Once serverless is fully deprecated, move all account scoped lambda calls to go via fetchAccountScopedLambdaApi.ts methods
+ * TODO: Then refactor this to be a generic base method that fetchHrmApi, fetchResourcesApi and fetchAccountScopedApi all call to add the token to the request
  */
-const fetchProtectedApi = async (
+export const fetchProtectedApi = async (
   endpoint: string,
   body: Record<string, any> = {},
   allOptions?: FetchOptions & { useTwilioLambda?: boolean; useJsonEncode?: boolean },
@@ -48,6 +49,7 @@ const fetchProtectedApi = async (
 
   // Adding the token to the body is for backwards compatibility only
   // Once serverless is fully deprecated and all account scoped lambdas are past v2.65.x it can be removed
+  // Also, support for form encoded payloads can probably be removed once serverless is deprecated too, since account-scoped lambda supports JSON request bodies on all endpoints
   const { contentType, encodedBody } = useJsonEncode
     ? { contentType: 'application/json', encodedBody: JSON.stringify({ ...body, Token: token }) }
     : {
@@ -67,36 +69,6 @@ const fetchProtectedApi = async (
   };
   try {
     return await fetchApi(new URL(useTwilioLambda ? accountScopedLambdaBaseUrl : serverlessBaseUrl), endpoint, options);
-  } catch (error) {
-    if (error instanceof ApiError) {
-      const message = error.response?.status === 403 ? 'Server responded with 403 status (Forbidden)' : error.message;
-      throw new ProtectedApiError(message, { response: error.response, body: error.body }, error);
-    } else throw error;
-  }
-};
-
-// eslint-disable-next-line import/no-unused-modules
-export const postToAccountScopedLambda = async (
-  endpoint: string,
-  body: Record<string, any> = {},
-  allOptions?: FetchOptions & { useJsonEncode?: boolean },
-) => fetchProtectedApi(endpoint, body, { ...(allOptions ?? {}), useTwilioLambda: true });
-
-export const getFromAccountScopedLambda = async (endpoint: string, fetchOptions?: FetchOptions) => {
-  const { accountScopedLambdaBaseUrl } = getHrmConfig();
-  const token = getValidToken();
-  if (token instanceof Error) throw new ApiError(`Aborting request due to token issue: ${token.message}`, {}, token);
-
-  const options: RequestInit = {
-    method: 'GET',
-    ...(fetchOptions ?? {}),
-    headers: {
-      Authorization: `Bearer ${token}`,
-      ...fetchOptions.headers,
-    },
-  };
-  try {
-    return await fetchApi(new URL(accountScopedLambdaBaseUrl), endpoint, options);
   } catch (error) {
     if (error instanceof ApiError) {
       const message = error.response?.status === 403 ? 'Server responded with 403 status (Forbidden)' : error.message;
